+++ tests/test_actions.py
@@ -1,7 +1,7 @@
 import os
 import tempfile
-
-from ccbr_actions.actions import use_github_action
+from ccbr_tools.shell import exec_in_context
+from ccbr_actions.actions import use_github_action, set_output
 
 
 def test_use_github_action():
@@ -10,3 +10,9 @@ def test_use_github_action():
             name="docs-mkdocs", save_as=os.path.join(tmpdir, "docs-mkdocs.yml")
         )
         assert os.path.exists(os.path.join(tmpdir, "docs-mkdocs.yml"))
+
+
+def test_set_output():
+    assert exec_in_context(set_output, "NAME", "VALUE", environ="ABC").startswith(
+        "::set-output name=NAME::VALUE\n"
+    )